delete field_weather_util, finish syncing field_weather_effects
