--- include/field_weather_util.h
@@ -1,6 +1,0 @@
-#ifndef GUARD_FIELD_WEATHER_UTIL_H
-#define GUARD_FIELD_WEATHER_UTIL_H
-
-void SetWeather(u8 weather);
-
-#endif //GUARD_FIELD_WEATHER_UTIL_H
