+++ src/debug.c
@@ -24,8 +24,7 @@
 #include "field_message_box.h"
 #include "field_screen_effect.h"
 #include "field_weather.h"
-#include "field_weather_util.h"
-// #include "international_string_util.h"
+#include "international_string_util.h"
 #include "item.h"
 #include "item_icon.h"
 #include "list_menu.h"
